requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
torch
torchvision
opencv-python
Flask
Pillow